requirements.txt
pyTelegramBotAPI==4.22.0
python-pptx==1.0.2
ollama==0.4.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.4.0